feat: watch cayos to recompile them, and rerender relevant pages
Also cleaned up Cayo component related files and names, so it's more
obviously compiling Cayo component instances, rather than normal
components.
