+++ lib/cli/index.js
@@ -107,15 +107,12 @@ async function run(command) {
     // Create a fresh cayo folder for this run
     cleanCayoPath(cayoConfig.cayoPath);
 
-    // Compile internal cayo component
-    await compile.cayoComponent(cayoConfig);
     // Compile layout template that will be used by pages
     const Layout = await compile.layout(cayoConfig);
     _cayo.layout = await Layout.render(true);
     
     // Compile & Render all user page and their dependent component files
     await compile.pages(null, _cayo, cayoConfig);
-    
 
     // A side effect of compiling all pages is having a full depedency tree
     // so we can up
@@ -125,23 +122,9 @@ async function run(command) {
       writePageFiles(page, cayoConfig.cayoPath, cayoConfig);
     }
 
-    await compile.components(null, _cayo, cayoConfig);
+    // TODO: decide: Need to pass null, or just remove that arg?
+    await compile.cayos(null, _cayo, cayoConfig);
 
-    // for (const name in _cayo.stats.cayoComponents) {
-    //   // TODO: this bundle stuff needs to go within
-    //   const { src } = _cayo.stats.cayoComponents[name];
-    //   const component = buildBundle(src, cayoConfig);
-    //   writeComponentFile(
-    //     { ...component, name }, 
-    //     src,
-    //     cayoConfig.cayoPath
-    //   );
-    // }
-
-
-    // let temp = _cayo.pages.find(page => page.name === 'index');
-    // console.log(temp.name, temp.dependencies);
-    
     // for debugging
     debugStats(_cayo.stats, cayoConfig);
