+++ lib/cli/watch.js
@@ -31,7 +31,7 @@ export function watch(config, _cayo) {
 
   const logChange = (type, filepath = '') => {
     if (filepath !== '') {
-      if (type === 'component') {
+      if (type === 'component' || 'cayo') {
         filepath = filepath.replace(config.components, '');
       } else if (type === 'page') {
         filepath = filepath.replace(config.pages, '');
@@ -69,49 +69,33 @@ export function watch(config, _cayo) {
           console.log('cayoComponents', _cayo.stats.cayoComponents);
           page.render(_cayo.stats.cayoComponents, { load: true });
         }
-        
-      
+
       } else if (filepath.startsWith(config.pages)) {
         logChange('page', filepath);
         // Handle Pages
         await handlePage(filepath, _cayo, config);
       
-      // Handle Cayo Components
+      // Handle Cayos
       } else if (filepath.includes(`.${config.cayoComponentInfix}`)) {
-        logChange('cayo component', filepath);
-        // compile.components([filepath], stats, config);
-        // find out which pages are affected and recompile/render those
-        // 2. render all affected pages
-
-        // getCayoComponents(config)
-        //   .then((components) => {
-        //     let componentModule = Object.entries(components).find(([, { modulePath }]) => modulePath === filepath);
-        //     handleCayoComponent(componentModule[0], componentModule[1].modulePath, config);
-        //   })
+        logChange('cayo', filepath);
+        await handleCayo(filepath, _cayo, config);
 
       // Handle Components
       } else if (filepath.startsWith(config.components)) {
         logChange('component', filepath);
         await handleComponent(filepath, _cayo, config);
-        // compile.components([filepath], stats, config);
-        // find out which pages are affected and recompile/render those
-        // 2. render all affected pages
-
-        // getPages(config)
-        //   .then(() => {
-        //     prerenderPages(config);
-        //   })
-  
-      // TODO: watch component changes
-      // } else if (componentFileChanged) {
-      // find out which pages are affected
-      // find out which components are affected (imports)?
       } else {
+        // TODO: what happens?
         // prerenderPages(config);
       }
     }
   });
   // watcher.close();
+}
+
+async function handleCayo(filepath, _cayo, config) {
+  await compile.cayos(null, _cayo, config);
+  handleComponent(filepath, _cayo, config);
 }
 
 async function handleComponent(filepath, _cayo, config) {
