+++ lib/core/compile/index.js
@@ -1,4 +1,3 @@
-export { compileCayoComponent as cayoComponent } from './cayoComponent.js';
 export { compileLayout as layout } from './layout.js';
 export { compilePages as pages } from './pages.js';
-export { compileComponents as components } from './components.js';
+export { compileCayos as cayos } from './cayos.js';
