+++ lib/core/compile/layout.js
@@ -12,6 +12,8 @@ export async function compileLayout(config) {
   let sourcePath = path.resolve(src, `${templateName}.svelte`);
   try {
     const source = await fs.readFile(sourcePath, 'utf8');
+    // TODO: support user-defined preprocessors for layout too?
+    // TODO: change this to use bundle.js
     const template = await preprocess(source, sveltePreprocess(), { filename });
     const {
       js,
